Add ability to create a new game from GameList
